Duke class: Extract functions into classes
Extracted the following from the duke class:
* loading and saving of tasks in the data file (Storage.java)
* making sense of the user command (Parser.java)
* the taskList, and operations that manipulate it (TaskList.java)

This was done to make the code more object oriented
