+++ data/duke.txt
@@ -1,3 +1,4 @@
 T | 0 | wows
 D | 1 | magic | 23/01/2024 1234
 E | 0 | sweet | 05/03/2024 1200 | 06/04/2024 2359
+D | 1 | max | 01/11/2024 1234
