+++ src/main/java/Duke.java
@@ -1,4 +1,5 @@
 import java.io.FileNotFoundException;
+import java.util.Arrays;
 import java.util.Scanner;
 import java.io.File;
 import java.io.FileWriter;
@@ -15,8 +16,7 @@ public class Duke {
     static String goodbyeMessage = horzLine +
             "\nBye. Hope to see you again soon!\n"
             + horzLine;
-    final static String dataPath = System.getProperty("user.dir") + "/data";
-    final static String fileName = "duke.txt";
+
     public static void printWithLines(String message) {
         System.out.println(horzLine);
         System.out.println(message);
@@ -24,14 +24,18 @@ public class Duke {
     }
 
     private Ui ui;
-    private ArrayList<Task> taskList = new ArrayList<>(100);
+    private Storage storage;
+    private TaskList tasks;
+    private Parser parser;
 
     public Duke() {
-        ui =new Ui();
+        ui = new Ui();
+        storage = new Storage();
+        tasks = new TaskList(storage.loadData());
+        parser = new Parser();
     }
 
     public void run() {
-        loadData();
         System.out.println(greetingMessage);
 
         while(!ui.getUserInput().equalsIgnoreCase("bye")) {
@@ -42,7 +46,7 @@ public class Duke {
                 break;
             }
 
-            String[] userMessageArr = getCommand(userMessage);
+            String[] userMessageArr = parser.getCommand(userMessage);
             String userCmd = userMessageArr[0];
             String cmdDetails = userMessageArr[1];
 
@@ -51,7 +55,7 @@ public class Duke {
                 AcceptedCmds testCommand = AcceptedCmds.valueOf(userCmd.toLowerCase());
             } catch (IllegalArgumentException e) {
                 printWithLines("Please enter a valid command\nThe list of valid commands are as follows:\n" +
-                        java.util.Arrays.asList(AcceptedCmds.values()));
+                        Arrays.asList(AcceptedCmds.values()));
                 continue;
             }
 
@@ -60,7 +64,7 @@ public class Duke {
             } else if (userCmd.equalsIgnoreCase("mark") ||
                     userCmd.equalsIgnoreCase("unmark")) {
                 try {
-                    String possInteger = getCmdDetails(userCmd, cmdDetails);
+                    String possInteger = parser.getCmdDetails(userCmd, cmdDetails);
                     int taskIndex = Integer.valueOf(possInteger);
 
                     if (userCmd.equalsIgnoreCase("unmark")) {
@@ -71,37 +75,37 @@ public class Duke {
                 } catch (DukeException e) {
                     printWithLines(e.getMessage());
                 } catch (NumberFormatException e) {
-                    printWithLines((taskList.size() != 0 ?
-                            "Invalid input type\nEnter a number between 1 and " + taskList.size() :
+                    printWithLines((tasks.size() != 0 ?
+                            "Invalid input type\nEnter a number between 1 and " + tasks.size() :
                             "Invalid input type\nCan't mark or unmark either cause the list is empty"));
                 }
             } else if (userCmd.equalsIgnoreCase("delete")) {
                 try {
-                    String possInteger = getCmdDetails(userCmd, cmdDetails);
+                    String possInteger = parser.getCmdDetails(userCmd, cmdDetails);
                     int taskIndex = Integer.valueOf(possInteger);
-                    delete(taskIndex);
+                    tasks.delete(taskIndex);
                 } catch (DukeException e) {
                     printWithLines(e.getMessage());
                 } catch (NumberFormatException e) {
-                    printWithLines((taskList.size() != 0 ?
-                            "Invalid input type\nEnter a number between 1 and " + taskList.size() :
+                    printWithLines((tasks.size() != 0 ?
+                            "Invalid input type\nEnter a number between 1 and " + tasks.size() :
                             "Invalid input type\nCan't mark or unmark either cause the list is empty"));
                 }
             } else if (userCmd.equalsIgnoreCase("todo")) {
                 try {
-                    String possToDo = getCmdDetails(userCmd, cmdDetails);
+                    String possToDo = parser.getCmdDetails(userCmd, cmdDetails);
                     ToDo newToDo = new ToDo(possToDo);
-                    addTask(newToDo);
+                    tasks.addTask(newToDo);
                 } catch (DukeException e) {
                     printWithLines(e.getMessage());
                 }
             } else if (userCmd.equalsIgnoreCase("deadline")) {
                 try {
-                    String possDLDetails = getCmdDetails(userCmd, cmdDetails);
+                    String possDLDetails = parser.getCmdDetails(userCmd, cmdDetails);
                     String[] splitDetails = possDLDetails.toLowerCase().split("/by ", 2);
                     try {
                         Deadline newDL = new Deadline(splitDetails[0], splitDetails[1]);
-                        addTask(newDL);
+                        tasks.addTask(newDL);
                     } catch (ArrayIndexOutOfBoundsException e) {
                         printWithLines("After entering the deadline task name,\n" +
                                 "add '/by' followed by your desired deadline");
@@ -112,14 +116,14 @@ public class Duke {
 
             } else if (userCmd.equalsIgnoreCase("event")){
                 try {
-                    String possEventDetails = getCmdDetails(userCmd, cmdDetails);
+                    String possEventDetails = parser.getCmdDetails(userCmd, cmdDetails);
                     String[] splitDetails = possEventDetails.split("/from ", 2);
                     String[] secondSplitDetails = new String[2];
                     try {
                         secondSplitDetails = splitDetails[1].split("/to ", 2);
                         try {
                             Event newEvent = new Event(splitDetails[0], secondSplitDetails[0], secondSplitDetails[1]);
-                            addTask(newEvent);
+                            tasks.addTask(newEvent);
                         } catch (ArrayIndexOutOfBoundsException e) {
                             printWithLines("After entering your desired start time,\n" +
                                     "add '/to' followed by your desired end time");
@@ -132,177 +136,54 @@ public class Duke {
                     printWithLines(e.getMessage());
                 }
             }
-            saveData();
+            storage.saveData(tasks);
         }
         System.out.println(goodbyeMessage);
     }
 
-    public void loadData() {
-        File directory = new File(dataPath);
-        File file = new File(dataPath + "/" + fileName);
-        ArrayList<Task> tempTaskStorage = new ArrayList<>(100);
 
-        // check if data directory and file exist already
-        if (!directory.exists()){
-            directory.mkdir();
-            try {
-                file.createNewFile();
-                System.out.println("First initialization, creating new save file...");
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        } else if (!file.exists()) {
-            try {
-                file.createNewFile();
-                System.out.println("Data file missing, creating new save file...");
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        } else {
-            try {
-                Scanner s = new Scanner(file);
-                while (s.hasNextLine()) {
-                    String newEntry = s.nextLine();
-                    String[] entryDetails = newEntry.split(" \\| ");
-                    Task newTask;
-                    switch (entryDetails[0]) {
-                        case "T":
-                            newTask = new ToDo(entryDetails[1], entryDetails[2]);
-                            tempTaskStorage.add(newTask);
-                            break;
-                        case "D":
-                            newTask = new Deadline(entryDetails[1], entryDetails[2], entryDetails[3]);
-                            tempTaskStorage.add(newTask);
-                            break;
-                        case "E":
-                            newTask = new Event(entryDetails[1], entryDetails[2], entryDetails[3], entryDetails[4]);
-                            tempTaskStorage.add(newTask);
-                            break;
-                    }
-                }
-                taskList.addAll(tempTaskStorage);
-            } catch (DukeException | ArrayIndexOutOfBoundsException e) {
-                try {
-                    file.delete();
-                    file.createNewFile();
-                    System.out.println("Data file corrupted, creating new save file...");
-                } catch (IOException ioe) {
-                    ioe.printStackTrace();
-                }
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
-            }
-        }
-    }
 
-    public void saveData() {
-        File directory = new File(dataPath);
-        File file = new File(dataPath + "/" + fileName);
-        // reset file
-        try {
-            file.delete();
-            file.createNewFile();
-            FileWriter writer = new FileWriter(file);
-            for (Task tsk: taskList) {
-                if (tsk instanceof ToDo) {
-                    writer.write("T | " + tsk.getStatusAsNum() + " | " + tsk.getDescription()
-                            + System.lineSeparator());
-                } else if (tsk instanceof Deadline) {
-                    writer.write("D | " + tsk.getStatusAsNum() + " | " + tsk.getDescription()
-                            + " | " + ((Deadline) tsk).getBy()
-                            + System.lineSeparator());
-                } else if (tsk instanceof Event) {
-                    writer.write("E | " + tsk.getStatusAsNum() + " | " + tsk.getDescription()
-                            + " | " + ((Event) tsk).getFrom() + " | " + ((Event) tsk).getTo()
-                            + System.lineSeparator());
-                }
-            }
-            writer.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 
-    public void addTask(Task newTask) {
-        taskList.add(newTask);
-        printWithLines("Got it. I've added this task:\n   " + newTask.toString() +
-                "\nNow you have " + taskList.size() + (taskList.size() > 1 ? " tasks ": " task ") +
-                "in the list.");
-    }
 
     public void list() {
-        if (taskList.size() == 0) {
+        if (tasks.size() == 0) {
             printWithLines("There's nothing in your list so far");
             return;
         }
         System.out.println(horzLine);
         System.out.println("Here are the tasks in your list:");
-        for (int i = 0; i < taskList.size(); i++) {
+        for (int i = 0; i < tasks.size(); i++) {
             int j = i + 1;
-            System.out.println(j + ". " + taskList.get(i).toString());
+            System.out.println(j + ". " + tasks.get(i).toString());
         }
         System.out.println(horzLine);
     }
 
-    public void delete(int index) throws DukeException {
-        if (taskList.size() == 0) {
-            throw new DukeException("Nothing is in the list yet");
-        }
-        if (index < 1 || index > taskList.size()) {
-            throw new DukeException("Please enter a number between 1 and " + taskList.size());
-        }
-        Task tempTask = taskList.get(index - 1);
-        taskList.remove(index - 1);
-        printWithLines("Noted. I've removed this task:\n   " + tempTask.toString() +
-                "\nNow you have " + taskList.size() + (taskList.size() > 1 ? " tasks ": " task ") +
-                "in the list.");
-    }
+
 
     public void markDone(int index) throws DukeException {
-        if (taskList.size() == 0) {
+        if (tasks.size() == 0) {
             throw new DukeException("Nothing is in the list yet");
         }
-        if (index < 1 || index > taskList.size()) {
-            throw new DukeException("Please enter a number between 1 and " + taskList.size());
+        if (index < 1 || index > tasks.size()) {
+            throw new DukeException("Please enter a number between 1 and " + tasks.size());
         }
-        taskList.get(index - 1).markAsDone();
-        printWithLines("Nice! I've marked this task as done:\n  " + taskList.get(index - 1).toString());
+        tasks.get(index - 1).markAsDone();
+        printWithLines("Nice! I've marked this task as done:\n  " + tasks.get(index - 1).toString());
     }
 
     public void markNotDone(int index) throws DukeException {
-        if (taskList.size() == 0) {
+        if (tasks.size() == 0) {
             throw new DukeException("Nothing is in the list yet");
         }
-        if (index < 1 || index > taskList.size()) {
-            throw new DukeException("Please enter a number between 1 and " + taskList.size());
+        if (index < 1 || index > tasks.size()) {
+            throw new DukeException("Please enter a number between 1 and " + tasks.size());
         }
-        taskList.get(index - 1).markAsUndone();
-        printWithLines("OK, I've marked this task as not done yet:\n  " + taskList.get(index - 1).toString());
+        tasks.get(index - 1).markAsUndone();
+        printWithLines("OK, I've marked this task as not done yet:\n  " + tasks.get(index - 1).toString());
     }
 
-    public String[] getCommand(String userMessage) {
-        String[] result = new String[2];
-        Boolean foundSplit = false;
-        for (int i = 0; i < userMessage.length(); i++) {
-            if (userMessage.charAt(i) == ' ') {
-                result[0] = userMessage.substring(0, i);
-                result[1] = userMessage.substring(i + 1, userMessage.length());
-                foundSplit = true;
-                break;
-            }
-        }
-        if (!foundSplit) {
-            result[0] = userMessage;
-        }
-        return result;
-    }
 
-    public String getCmdDetails(String cmd, String details) throws DukeException {
-        if (details == null || details.trim().length() == 0) {
-            throw new DukeException("Please enter a description for the " + cmd + " command");
-        }
-        return details.trim();
-    }
     public static void main(String[] args) {
         new Duke().run();
     }
